feat: refactor serde
