+++ manta-trusted-setup/Cargo.toml
@@ -57,6 +57,7 @@ manta-util = { path = "../manta-util", default-features = false, features = ["se
 rand_chacha = { version = "0.3.1", default-features = false }
 reqwest = {version = "0.11.11", features = ["json"]}
 serde = { version = "1.0", default-features = false, features = ["derive"]  }
+serde_arrays = "0.1.0"
 serde_json = {version = "1.0", default-features = false, features = ["alloc"]}
 serde_with = { version = "1.14.0", default-features = false }
 take_mut = "0.2.2" # TODO: review this dependency
